Bullets And Walls 1.0
The Military saw your Car collision project and they were interested in assigning you another project. They have given you $1000 to test the reliability of walls when bullets hit them.

They test the effectiveness of walls against bullets, amongst other things. All walls are made of concrete material. If the damage to the wall from the bullet exceeds a certain value, the material used to make the wall is rejected.